requirements.txt
streamlit
torch
torchvision
Pillow
numpy
matplotlib
scikit-learn
opencv-python
reportlab

